+++ core/src/commonMain/kotlin/com/boswelja/ephemeris/core/model/YearMonth.kt
@@ -19,7 +19,7 @@ public data class YearMonth(
 public val LocalDate.yearMonth: YearMonth
     get() = YearMonth(year, month)
 
-public fun YearMonth.plus(months: Long): YearMonth {
+public fun YearMonth.plus(months: Int): YearMonth {
     val monthDelta = month.ordinal + months
     val yearDelta = if (monthDelta < 0) {
         (monthDelta / allMonths.size) - 1
@@ -27,14 +27,14 @@ public fun YearMonth.plus(months: Long): YearMonth {
         (monthDelta / allMonths.size)
     }
     return copy(
-        year = (yearDelta + year).toInt(),
+        year = (yearDelta + year),
         month = month.plusMonths(months)
     )
 }
 
 public fun YearMonth.until(other: YearMonth): Int {
     val yearDelta = year - other.year
-    val monthDelta = month.plusMonths((-1 * other.month.number).toLong()).number
+    val monthDelta = month.plusMonths(-1 * other.month.number).number
     return (yearDelta * allMonths.size) + monthDelta
 }
